services/heartbeat/requirements.txt
# HeartBeat Service Runtime Dependencies

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Password hashing (API key management)
bcrypt>=4.0.0

# HTTP Client (inter-service communication, health checks)
httpx>=0.25.0

# Async support
anyio>=4.0.0

# Logging
python-json-logger>=2.0.0

# Configuration
python-dotenv>=1.0.0

# Observability (Phase 2)
prometheus_client>=0.20.0

# Auth (Part 4 -- JWT, Ed25519)
authlib>=1.3.0
cryptography>=41.0.0

# UUID v7 (time-sortable identifiers)
uuid6>=2.0.0

# PostgreSQL (March 2026 -- replaces SQLite for server-side)
psycopg2-binary>=2.9.0

# Legacy SQLite (dev only, will be removed)
# sqlcipher3-binary>=0.5.0
